requirements_dev.txt
--index-url https://pypi.python.org/simple/
-e .
pytest-flake8==0.2.0
pytest-watch==4.0.0
pytest-benchmark==3.0.0
pytest-cov==2.2.1
pytest-mock==0.11.0
